SCRUM-146 feat: change types
