+++ src/app/page.tsx
@@ -1,11 +1,7 @@
-import type {
-  GetCommentsResponse,
-  PublicPostsResponse,
-  UsersCountResponse,
-} from '@/src/shared/model/api/types'
-
 import { AuthWrapper } from '../features/authWrapper/AuthWrapper'
-import { Post } from '@/src/entities/post/types'
+import { GetCommentsResponse } from '@/src/entities/comments/types'
+import { Post, PublicPostsResponse } from '@/src/entities/post/types'
+import { UsersCountResponse } from '@/src/entities/users/types'
 import { SearchParams, getUserComments, getUserPost } from '@/src/widgets/profile/getPublicProfile'
 import { PublicFeed } from '@/src/widgets/publicFeed/PublicFeed'
 import { getUsersCount, getUsersPosts } from '@/src/widgets/publicFeed/getPublicPosts'
