+++ src/features/croppingPhoto/CroppingPhotoProfile/CroppingPhotoProfile.tsx
@@ -3,9 +3,9 @@ import { useState } from 'react'
 import Cropper from 'react-easy-crop'
 import { useDispatch } from 'react-redux'
 
+import { CustomerError } from '@/src/entities/errors/types'
 import { applyCropToAllPhotos } from '@/src/features/croppingPhoto/applyCropToAllPhotos'
 import { PhotoSettings } from '@/src/features/croppingPhoto/types'
-import { CustomerError } from '@/src/shared/model/api/types'
 import { useUpdateUserAvatarMutation } from '@/src/shared/model/api/usersApi'
 import { setIsPhotoModalOpen } from '@/src/shared/model/slices/modalSlice'
 import { Alerts } from '@/src/shared/ui/alerts/Alerts'
