+++ src/features/forgotPassword/ForgotPasswordCard.tsx
@@ -4,9 +4,9 @@ import { ChangeEvent, useRef, useState } from 'react'
 import ReCAPTCHA from 'react-google-recaptcha'
 import { SubmitHandler, useForm } from 'react-hook-form'
 
+import { CustomerError } from '@/src/entities/errors/types'
 import { FormType, schema } from '@/src/features/forgotPassword/validators'
 import { usePasswordRecoveryMutation } from '@/src/shared/model/api/authApi'
-import { CustomerError } from '@/src/shared/model/api/types'
 import { Button } from '@/src/shared/ui/button/Button'
 import { Card } from '@/src/shared/ui/card/Card'
 import { Dialog } from '@/src/shared/ui/dialog/Dialog'
