+++ src/features/login/Login.tsx
@@ -3,10 +3,10 @@
 import React from 'react'
 import { SubmitHandler, useForm } from 'react-hook-form'
 
+import { LoginError } from '@/src/entities/errors/types'
 import { FormType, schema } from '@/src/features/login/validators'
 import { AuthRoutes } from '@/src/shared/lib/constants/routing'
 import { useLoginMutation } from '@/src/shared/model/api/authApi'
-import { LoginError } from '@/src/shared/model/api/types'
 import { Button } from '@/src/shared/ui/button/Button'
 import { Card } from '@/src/shared/ui/card/Card'
 import { Input } from '@/src/shared/ui/input'
