+++ src/features/register/Register.tsx
@@ -3,10 +3,10 @@
 import { useRef, useState } from 'react'
 import { SubmitHandler, useForm } from 'react-hook-form'
 
+import { CustomerError } from '@/src/entities/errors/types'
 import { FormType, schema } from '@/src/features/register/validators'
 import { AuthRoutes } from '@/src/shared/lib/constants/routing'
 import { useRegistrationMutation } from '@/src/shared/model/api/authApi'
-import { CustomerError } from '@/src/shared/model/api/types'
 import { Button } from '@/src/shared/ui/button/Button'
 import { Card } from '@/src/shared/ui/card/Card'
 import { CheckBox } from '@/src/shared/ui/checkbox/CheckBox'
