+++ src/shared/hooks/useAddAnswer.ts
@@ -1,6 +1,8 @@
 import { useState } from 'react'
+
+import { AnswersComment } from '@/src/entities/comments/types'
+import { CustomerError } from '@/src/entities/errors/types'
 import { useAddAnswerMutation } from '@/src/shared/model/api/postsApi'
-import { AnswersComment, CustomerError } from '@/src/shared/model/api/types'
 import { setAppError } from '@/src/shared/model/slices/appSlice'
 import { useAppDispatch } from '@/src/shared/model/store/store'
 
@@ -15,19 +17,23 @@ export const useAddAnswer = (postId: number, commentId: number) => {
   }
 
   const handleSubmit = async (): Promise<AnswersComment | null> => {
-    if (!commentText.trim()) return null
+    if (!commentText.trim()) {
+      return null
+    }
 
     try {
-      const result = await addAnswer({ postId, commentId, content: commentText }).unwrap()
+      const result = await addAnswer({ commentId, content: commentText, postId }).unwrap()
 
       setCommentText('')
+
       return result
     } catch (err) {
       const error = err as CustomerError
       const errorMessage =
-        error.data?.messages[0].message || error.data?.error || 'Failed to add comment'
+        error.data?.messages[0].message || error.data?.error || 'Failed to add comments'
 
       dispatch(setAppError({ error: errorMessage }))
+
       return null
     }
   }
