+++ src/shared/model/api/searchApi.ts
@@ -1,5 +1,5 @@
+import { GetSearchUserArgs, GetSearchUserResponse } from '@/src/entities/search/types'
 import { baseApi } from '@/src/shared/model/api/baseApi'
-import { GetSearchUserArgs, GetSearchUserResponse } from '@/src/shared/model/api/types'
 
 export const searchApi = baseApi.injectEndpoints({
   endpoints: builder => ({
