+++ src/shared/model/api/socketApi.ts
@@ -1,5 +1,5 @@
+import { Notifications } from '@/src/entities/notifications/types'
 import { notificationsApi } from '@/src/shared/model/api/notificationsApi'
-import { Notifications } from '@/src/shared/model/api/types'
 import { AppDispatch } from '@/src/shared/model/store/store'
 import { Socket, io } from 'socket.io-client'
