+++ src/shared/model/api/usersApi.ts
@@ -1,4 +1,4 @@
-import { Profile, ProfileByUserName, PublicProfileTypes } from '@/src/entities/user/types'
+import { Profile, ProfileByUserName, PublicProfileTypes } from '@/src/entities/users/types'
 import { baseApi } from '@/src/shared/model/api/baseApi'
 
 export const usersApi = baseApi.injectEndpoints({
