+++ src/widgets/answerItem/AnswerItem.tsx
@@ -1,8 +1,8 @@
+import { AnswersComment } from '@/src/entities/comments/types'
 import Heart from '@/src/shared/assets/componentsIcons/Heart'
 import HeartOutline from '@/src/shared/assets/componentsIcons/HeartOutline'
 import { timeSince } from '@/src/shared/lib/timeSince'
 import { useGetAnswerLikesQuery } from '@/src/shared/model/api/postsApi'
-import { AnswersComment } from '@/src/shared/model/api/types'
 import { AvatarBox } from '@/src/shared/ui/avatar/AvatarBox'
 import { Button } from '@/src/shared/ui/button/Button'
 import { Typography } from '@/src/shared/ui/typography/Typography'
