+++ src/widgets/commentItemContainer/CommentItemContainer.tsx
@@ -1,5 +1,6 @@
+import { AnswersComment, Comment } from '@/src/entities/comments/types'
+import { LikeStatus } from '@/src/entities/likes/types'
 import { useGetCommentLikesQuery } from '@/src/shared/model/api/postsApi'
-import { AnswersComment, Comment, LikeStatus } from '@/src/shared/model/api/types'
 import { CommentItem } from '@/src/widgets/commentItem/CommentItem'
 
 type Props = {
